to not display the logs

+++ components/frontend.sh
@@ -22,7 +22,17 @@
 
 yum install nginx -y
 
+if [ $? -ne 0 ]; then
+  echo "Nginx Installation is Failed"
+  exit
+fi
+
 curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zip"
+
+if [ $? -ne 0 ]; then
+  echo "Download front-end failed"
+  exit
+fi
 
 cd /usr/share/nginx/html
 rm -rf *
